requirements.txt
streamlit==1.13.0
pandas<1.5.0
numpy==1.23.3
matplotlib>3.0.1
jinja2==3.1.2
watchdog==2.1.9
coverage==6.5.0
pytest-cov==4.0.0
plotly==5.10.0
graphviz==0.20.1
plotnine==0.10.1
altair==4.2.0
scikit-learn==1.1.2
seaborn==0.11.2
Pillow==9.2.0
